+++ test/dummy/app/models/folder.rb
@@ -1,3 +1,0 @@
-class Folder < ApplicationRecord
-  recording_studio_recordable label: "Folder", root: false, allowed_parent_types: [ "Workspace", "Folder" ]
-end