+++ test/dummy/config/initializers/recording_studio_billing.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+RecordingStudioBilling.configure do |config|
+  config.provider = :stripe
+end